test-runner [guile]: module system workaround
Guile's module system makes nested modules available in the namespace
of their containing module, and this binding overrides any imported
names. In `(spells test-runner)', this lead to the `environment'
binding being overridden.  This might arguably be a bug in Guile's
module system implementation, but since it is easy to avoid (at least
in this case), we do so.